+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - More settings implemented.
 - Signal routing is now an exclusive operation, for practical reasons.
+- Granular synthesizer. FunUTD is now a dependency.
 
 ### Version 0.11